feat: 组织架构侧边栏 #1610 # Reviewed, transaction id: 6689
