+++ src/pages/src/http/organizationFiles.ts
@@ -70,3 +70,8 @@ export const addDepartment = (id: string, params: any) => http.post(`${prefix}/t
  * 更新租户部门
  */
 export const updateDepartment = (id: string, params: any) => http.put(`${prefix}/tenants/departments/${id}/`, params);
+
+/**
+ * 获取当前租户的协作租户信息
+ */
+export const getCollaboration = () => http.get(`${prefix}/collaborative-tenants/`);
