+++ src/pages/src/language/zh.json
@@ -452,5 +452,7 @@
   "变更": "变更",
   "添加子组织": "添加子组织",
   "重命名": "重命名",
+  "协同租户": "协同租户",
+  "协同配置": "协同配置",
   "新增": "新增"
 }
