+++ src/pages/src/store/app.ts
@@ -1,0 +1,11 @@
+import { defineStore } from 'pinia';
+import { computed, ref } from 'vue';
+
+
+export default defineStore('app', () => {
+  const currentOrg = ref({});
+
+  return {
+    currentOrg,
+  };
+});
