+++ src/pages/static/bk_icon_font/demo.html
@@ -121,6 +121,10 @@
           <span class="icon user-icon icon-app-store-fill"></span>
           <p class="icon-text">app-store-fill</p>
         </li>
+        <li class="icon-item" title="refresh">
+          <span class="icon user-icon icon-refresh"></span>
+          <p class="icon-text">refresh</p>
+        </li>
         <li class="icon-item" title="info-i">
           <span class="icon user-icon icon-info-i"></span>
           <p class="icon-text">info-i</p>
@@ -365,6 +369,14 @@
           <span class="icon user-icon icon-shezhi"></span>
           <p class="icon-text">shezhi</p>
         </li>
+        <li class="icon-item" title="tongbu">
+          <span class="icon user-icon icon-tongbu"></span>
+          <p class="icon-text">tongbu</p>
+        </li>
+        <li class="icon-item" title="help-document-fill">
+          <span class="icon user-icon icon-help-document-fill"></span>
+          <p class="icon-text">help-document-fill</p>
+        </li>
       </ul>
       <h3 class="describe-title">为什么使用</h3>
       <ul class="use-describe">
@@ -387,6 +399,12 @@
             <use xlink:href="#icon-app-store-fill"></use>
           </svg>
           <p class="icon-text">app-store-fill</p>
+        </li>
+        <li class="colorful-icon">
+          <svg class="icon svg-icon">
+            <use xlink:href="#icon-refresh"></use>
+          </svg>
+          <p class="icon-text">refresh</p>
         </li>
         <li class="colorful-icon">
           <svg class="icon svg-icon">
@@ -753,6 +771,18 @@
             <use xlink:href="#icon-shezhi"></use>
           </svg>
           <p class="icon-text">shezhi</p>
+        </li>
+        <li class="colorful-icon">
+          <svg class="icon svg-icon">
+            <use xlink:href="#icon-tongbu"></use>
+          </svg>
+          <p class="icon-text">tongbu</p>
+        </li>
+        <li class="colorful-icon">
+          <svg class="icon svg-icon">
+            <use xlink:href="#icon-help-document-fill"></use>
+          </svg>
+          <p class="icon-text">help-document-fill</p>
         </li>
       </ul>
       <h3 class="describe-title">为什么使用</h3>
