+++ src/pages/static/bk_icon_font/style.css
@@ -26,6 +26,9 @@ url("fonts/iconcool.eot?#iefix") format("embedded-opentype");
 .icon-app-store-fill:before {
 	content: "\e126";
 }
+.icon-refresh:before {
+	content: "\e146";
+}
 .icon-info-i:before {
 	content: "\e105";
 }
@@ -208,4 +211,10 @@ url("fonts/iconcool.eot?#iefix") format("embedded-opentype");
 }
 .icon-shezhi:before {
 	content: "\e145";
+}
+.icon-tongbu:before {
+	content: "\e147";
+}
+.icon-help-document-fill:before {
+	content: "\e148";
 }
